Implements handling of preview mode, allowing unit editors to bail out and be replaced with a different editor.
I think we should change the attribute name from `isPreview` to `preview`. That would comply with the Java convention, and the is* causes even more confusion in backbone where you feel like writing model.isPreview but have to write model.get('isPreview') though we could of course implement the former as a custom method.